+++ lib/pages/verification_screen.dart
@@ -1,0 +1,28 @@
+import 'package:flutter/material.dart';
+
+class VerificationScreen extends StatelessWidget {
+  final String email;
+
+  const VerificationScreen({Key? key, required this.email}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Email Verification'),
+      ),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Text(
+              'An email has been sent to $email for verification. Please check your inbox and follow the instructions.',
+              textAlign: TextAlign.center,
+              style: TextStyle(fontSize: 18),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
